# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Unified observability platform for AI agents, APIs, and infrastructure.",
        "title": "Keelwave API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/v1",
    "paths": {
...
        "/ingest/agent/steps": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Records one step in an agent loop. Server computes a SHA-256 fingerprint of tool_name + tool_input for loop detection and bumps the agent_tools registry when tool_name is present.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingest"
                ],
                "summary": "Append an agent step",
                "parameters": [
                    {
                        "description": "Step payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ingestAgentStepPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.ingestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
        "/projects/{projectID}/agent/runs/{runID}/loops": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns each tool_name + input fingerprint repeated \u003e= 2 times.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent"
                ],
                "summary": "Loop detection for an agent run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "projectID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Agent run UUID",
                        "name": "runID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 hint for chunk pruning",
                        "name": "at",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.LoopHit"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/projects/{projectID}/agent/runs/{runID}/steps": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent"
                ],
                "summary": "List steps for an agent run, ordered by step_index asc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "projectID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Agent run UUID",
                        "name": "runID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 hint for chunk pruning",
                        "name": "at",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default 1000, max 1000",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.AgentStep"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/projects/{projectID}/agent/steps/distribution": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent"
                ],
                "summary": "Step counts grouped by step_type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "projectID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RFC3339, default now - 30d",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339, default now",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.StepTypeCount"
                            }
                        }
                    }
                }
            }
        },
...
        "/projects/{projectID}/agent/tools/stats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Aggregates tool calls across every run in the window: call count, success/fail counts, success rate, and p95 latency per tool.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent"
                ],
                "summary": "Per-tool usage analytics for the calling project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "projectID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RFC3339, default now - 30d",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339, default now",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.ToolStat"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
...
        "main.ingestAgentStepPayload": {
            "type": "object",
            "required": [
                "agent_run_id",
                "step_type"
            ],
            "properties": {
                "agent_run_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string",
                    "maxLength": 100000
                },
                "cost_usd": {
                    "type": "number",
                    "minimum": 0
                },
                "metadata": {
                    "type": "object"
                },
                "step_index": {
                    "type": "integer",
                    "minimum": 0
                },
                "step_type": {
                    "type": "string",
                    "enum": [
                        "think",
                        "tool_call",
                        "tool_result",
                        "replan"
                    ]
                },
                "timestamp": {
                    "type": "string"
                },
                "tokens": {
                    "type": "integer",
                    "minimum": 0
                },
                "tool_input": {
                    "type": "object"
                },
                "tool_latency_ms": {
                    "type": "integer",
                    "minimum": 0
                },
                "tool_name": {
                    "type": "string",
                    "maxLength": 200
                },
                "tool_output": {
                    "type": "object"
                },
                "tool_success": {
                    "type": "boolean"
                }
            }
        },
...
        "main.ingestResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "store.AgentStep": {
            "type": "object",
            "properties": {
                "agent_run_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "cost_usd": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "input_fingerprint": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "metadata": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "project_id": {
                    "type": "string"
                },
                "step_index": {
                    "type": "integer"
                },
                "step_type": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "tokens": {
                    "type": "integer"
                },
                "tool_input": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "tool_latency_ms": {
                    "type": "integer"
                },
                "tool_name": {
                    "type": "string"
                },
                "tool_output": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "tool_success": {
                    "type": "boolean"
                }
            }
        },
...
        "store.LoopHit": {
            "type": "object",
            "properties": {
                "fingerprint": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "hits": {
                    "type": "integer"
                },
                "step_indices": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "tool_name": {
                    "type": "string"
                }
            }
        },
...
        "store.StepTypeCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "step_type": {
                    "type": "string"
                }
            }
        },
...
        "store.ToolStat": {
            "type": "object",
            "properties": {
                "call_count": {
                    "type": "integer"
                },
                "fail_count": {
                    "type": "integer"
                },
                "is_new": {
                    "type": "boolean"
                },
                "p50_latency_ms": {
                    "type": "integer"
                },
                "p95_latency_ms": {
                    "type": "integer"
                },
                "p99_latency_ms": {
                    "type": "integer"
                },
                "prev_call_count": {
                    "type": "integer"
                },
                "success_count": {
                    "type": "integer"
                },
                "success_rate": {
                    "type": "number"
                },
                "tool_name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Format: \"Bearer kw_\u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}